ctrl_s_master/requirements.txt
python-dotenv
bitwarden-sdk
pykeepass
cryptography
lxml
requests
pytest
argon2-cffi
dateutils